Upgrade to skrifa 0.42.0 and tiny-skia 0.12.0 with PDFPurr Git patch
- Add Git patch for PDFPurr to use tiny-skia 0.12.0 (PDFPurr 0.4.0 from crates.io uses 0.11.4)
- Fix skrifa_wrapper.rs: GlyphId::new to use u32 for skrifa 0.42.0
- Fix pdfpurr/mod.rs: Convert pixmap from tiny-skia 0.11.4 to 0.12.0 API (PremultipliedColorU8)
- Update documentation with new dependency versions and patch explanation
- Successfully build for ARM Kobo with upgraded dependencies

Dependencies:
- skrifa 0.42.0 (was 0.19.0)
- tiny-skia 0.12.0 (was 0.11.4)
- PDFPurr 0.4.0 from GitHub patch (for compatibility)

Phase 3 remains complete with MuPDF fully removed.

+++ crates/core/src/font/skrifa_wrapper.rs
@@ -173,7 +173,7 @@ impl Face {
 
         let hmtx = font.hmtx().ok();
         let advance_width = hmtx
-            .and_then(|h| h.advance(GlyphId::new(glyph_id as u16)))
+            .and_then(|h| h.advance(GlyphId::new(glyph_id as u32)))
             .unwrap_or(0) as i32;
 
         Ok(GlyphMetrics {
